Set reorder = FALSE in rowsum()
The previous default setting of reorder = TRUE seems like a mistake, and it in fact was the source of a major bug in a package I wrote. If the user is handing rowsum() a grouping vector, I would assume the user has set the grouping vector in the desired order by which they want to compute row-wise sums. If they wanted the grouping vector to be reordered prior to computing the row-wise sum, I expect that would be the exception rather than the rule, in which case they would set reorder = TRUE.